Added a Redux action to accept and display comments.
Created an ActionType.js file which contains the ADD_COMMENT type.
Created an addComment function in ActionCreators.js which will take in the new comment parameters, and export itself to comments.js where it can be updated to the state.
comments.js identifies the type of action and returns the new state
Added mapDispatchToProps in MainComponent so that addComment is available as a prop which can be passed on to DishDetailComponent.
Made the necessary props changes in DishDetail to accomodate the addComment function and call it on handleSubmit()